+++ core-ng/src/main/java/core/framework/impl/web/request/PathParams.java
@@ -1,6 +1,6 @@
 package core.framework.impl.web.request;
 
-import core.framework.api.util.ByteBuf;
+import core.framework.api.util.Encodings;
 import core.framework.api.util.Exceptions;
 import core.framework.api.util.Maps;
 import core.framework.api.web.exception.BadRequestException;
@@ -15,8 +15,11 @@ public final class PathParams {
 
     public void put(String name, String value) {
         if (value.length() == 0) throw new BadRequestException("path param must not be empty, name=" + name + ", value=" + value);
-        String previousValue = params.putIfAbsent(name, decodePathSegment(value));
-        if (previousValue != null) throw Exceptions.error("duplicated path variable found, name={}", name);
+        try {
+            params.put(name, Encodings.decodeURIComponent(value));
+        } catch (IllegalArgumentException e) {
+            throw new BadRequestException(e.getMessage(), BadRequestException.DEFAULT_ERROR_CODE, e);
+        }
     }
 
     public <T> T get(String name, Class<T> valueClass) {
@@ -51,34 +54,5 @@ public final class PathParams {
         } catch (NumberFormatException e) {
             throw new BadRequestException("failed to parse value to int, value=" + value, BadRequestException.DEFAULT_ERROR_CODE, e);
         }
-    }
-
-    // refer to http://www.ietf.org/rfc/rfc3986.txt, org.springframework.web.util.UriUtils#decode
-    String decodePathSegment(String path) {
-        int length = path.length();
-        int index = 0;
-        for (; index < length; index++) {
-            int ch = path.charAt(index);
-            if (ch == '%') break;
-        }
-        if (index == length) return path;
-        ByteBuf buffer = ByteBuf.newBuffer(length);
-        for (int i = 0; i < index; i++) buffer.put((byte) path.charAt(i));
-        for (; index < length; index++) {
-            int ch = path.charAt(index);
-            if (ch == '%') {
-                if ((index + 2) >= length) throw new BadRequestException("invalid path, value=" + path.substring(index));
-                char hex1 = path.charAt(index + 1);
-                char hex2 = path.charAt(index + 2);
-                int u = Character.digit(hex1, 16);
-                int l = Character.digit(hex2, 16);
-                if (u == -1 || l == -1) throw new BadRequestException("invalid path, value=" + path.substring(index));
-                buffer.put((byte) ((u << 4) + l));
-                index += 2;
-            } else {
-                buffer.put((byte) ch);
-            }
-        }
-        return buffer.text();
     }
 }