+++ core-ng/src/test/java/core/framework/impl/web/service/WebServiceClientTest.java
@@ -52,9 +52,9 @@ public class WebServiceClientTest {
         Assert.assertEquals("http://localhost/test", webServiceClient.serviceURL("/test", Maps.newHashMap()));
         Assert.assertEquals("http://localhost/test/", webServiceClient.serviceURL("/test/", Maps.newHashMap()));
 
-        Map<String, String> pathParams = Maps.newHashMap("id", "1");
-        Assert.assertEquals("http://localhost/test/1", webServiceClient.serviceURL("/test/:id(\\d+)", pathParams));
-        Assert.assertEquals("http://localhost/test/1", webServiceClient.serviceURL("/test/:id", pathParams));
+        Map<String, String> pathParams = Maps.newHashMap("id", "1+2");
+        Assert.assertEquals("http://localhost/test/1%2B2", webServiceClient.serviceURL("/test/:id(\\d+)", pathParams));
+        Assert.assertEquals("http://localhost/test/1%2B2", webServiceClient.serviceURL("/test/:id", pathParams));
     }
 
     @Test