+++ demo-site/src/main/java/app/web/Cookies.java
@@ -7,7 +7,7 @@ import java.time.Duration;
 /**
  * @author neo
  */
-public class CookieConstraints {
+public class Cookies {
     public static final CookieSpec TEST = new CookieSpec("test").httpOnly().sessionScope();
     public static final CookieSpec TEST1 = new CookieSpec("test1").httpOnly().maxAge(Duration.ofHours(2));
 }