+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Change log
 
+### 4.0.2 (1/18/2016)
+* web: removed URIBuilder, added Encodings.encodeURIComponent and decodeURIComponent
+* web: use URI query param encoding/decoding to set/get cookie (refered as URLEncoding in other place, e.g. jquery)
+
 ### 4.0.1 (1/13/2016)
 * schedule: support weekly and monthly trigger